refactor(frontend): move single-route components to route-private _components/
Relocate the seven single-consumer components whose only importer is one
route family into that route's _components/ folder, per the route-private
convention in the component reorg plan. Each is a git mv (rename detection
preserved) with the importing route's specifier rewritten to ./_components/.

  SearchResults                  -> routes/search/
  CreateFirstCorporateEntityPrompt -> routes/manufacturers/[slug]/
  ManufacturerCardGrid           -> routes/locations/[...path]/
  ModelRelationshipsList         -> routes/models/[slug]/
  NeedsReviewBanner              -> routes/titles/[slug]/
  CreateFirstModelPrompt         -> routes/titles/[slug]/
  RelatedTitlesSection           -> routes/titles/[slug]/

Defers the route-private candidates that conflict with feat/ssr-titles-faceting
(ThemeSwitcher, TitleFilterSidebar, ManufacturerActiveFilterChips).

Extend the move codemod to emit a relative path when a moved file's new home
is outside src/lib (route-private _components/); previously it left the
importer's $lib specifier unrewritten and broken.

Also rename the planned form/ folder to input/ and resolve several plan
inconsistencies (ExpandableSidebarList -> layout/page/sidebar/, stale
detail-section double-listings, ManufacturerCardGrid DRY follow-up).
