+++ frontend/src/routes/locations/[...path]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import AccordionSection from '$lib/components/AccordionSection.svelte';
-  import ManufacturerCardGrid from '$lib/components/ManufacturerCardGrid.svelte';
+  import ManufacturerCardGrid from './_components/ManufacturerCardGrid.svelte';
   import RichTextOverviewAccordion from '$lib/components/markdown/RichTextOverviewAccordion.svelte';
   import RichTextReferencesAccordion from '$lib/components/markdown/RichTextReferencesAccordion.svelte';
   import { createRichTextAccordionState } from '$lib/components/markdown/rich-text-accordion-state.svelte';
