+++ frontend/src/routes/locations/[...path]/_components/ManufacturerCardGrid.svelte
