+++ frontend/src/routes/models/[slug]/+page.svelte
@@ -8,7 +8,7 @@
   import RichTextOverviewAccordion from '$lib/components/markdown/RichTextOverviewAccordion.svelte';
   import RichTextReferencesAccordion from '$lib/components/markdown/RichTextReferencesAccordion.svelte';
   import { createRichTextAccordionState } from '$lib/components/markdown/rich-text-accordion-state.svelte';
-  import ModelRelationshipsList from '$lib/components/ModelRelationshipsList.svelte';
+  import ModelRelationshipsList from './_components/ModelRelationshipsList.svelte';
   import { modelEditActionContext } from '$lib/components/editors/edit-action-context';
   import { externalLinks } from '$lib/entities/external-links';
   import { model as modelInfo } from '$lib/entities/model';
