+++ frontend/src/routes/search/+page.svelte
@@ -2,7 +2,7 @@
   import { afterNavigate, replaceState } from '$app/navigation';
   import { onMount } from 'svelte';
   import SearchBox from '$lib/components/SearchBox.svelte';
-  import SearchResults from '$lib/components/SearchResults.svelte';
+  import SearchResults from './_components/SearchResults.svelte';
   import { SITE_TITLE } from '$lib/constants';
   import { resolve } from '$app/paths';
