+++ frontend/src/routes/titles/[slug]/_components/RelatedTitlesSection.svelte
